hotelplatformapi/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cloudinary==1.44.1
colorama==0.4.6
cryptography==45.0.5
Django==5.2.4
django-filter==25.1
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
pycparser==2.22
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
